move readinessprobe out
